+++ Sources/KeyboardKit/Feedback/HapticFeedback.swift
@@ -11,29 +11,44 @@ import Foundation
 /**
  This enum contains various haptic feedback types.
   
- You can call `prepare()` and `trigger()` on either the type
- or an instance, to prepare and trigger the desired feedback.
- 
- The feedback enum uses the static `player` to play feedback.
- You can replace this instance with a custom player, e.g. to
- mock functionality when writing tests.
- 
- Note that this enum is available on tvOS, where there is no
- haptic feedback. However,
+ You can call ``trigger()`` on any feedback value to trigger
+ the desired feedback, using the ``player``. You can replace
+ this player with any custom player if you need to.
+
+ Note that `macOS`, `tvOS` and `watchOS` resolves a disabled
+ ``HapticFeedbackPlayer`` by default, since the platforms do
+ not support haptic feedback. Feel free to replace it with a
+ player that does something meaningful on each platform.
 */
 public enum HapticFeedback: String, CaseIterable, Codable, Equatable, Identifiable {
     
     case
+
+    /// Represents feedback for an error event.
     error,
+
+    /// Represents feedback for a successful event.
     success,
+
+    /// Represents feedback for a warning event.
     warning,
-    
+
+
+    /// Represents light impact feedback.
     lightImpact,
+
+    /// Represents medium impact feedback.
     mediumImpact,
+
+    /// Represents heavy impact feedback.
     heavyImpact,
-    
+
+
+    /// Represents feedback when a selection changes.
     selectionChanged,
-    
+
+
+    /// Can be used to disable feedback.
     none
 }
 
@@ -43,15 +58,20 @@ public enum HapticFeedback: String, CaseIterable, Codable, Equatable, Identifiab
 public extension HapticFeedback {
     
     /**
-     The feedback's unique identifier.
+     The unique feedback identifier.
      */
     var id: String { rawValue }
-    
-    #if os(iOS)
-    static var player: HapticFeedbackPlayer = StandardHapticFeedbackPlayer.shared
-    #else
-    static var player: HapticFeedbackPlayer = DisabledHapticFeedbackPlayer()
-    #endif
+
+    /**
+     The player that will be used to trigger haptic feedback.
+     */
+    static var player: HapticFeedbackPlayer = {
+        #if os(iOS)
+        StandardHapticFeedbackPlayer.shared
+        #else
+        DisabledHapticFeedbackPlayer()
+        #endif
+    }()
     
     /**
      Prepare the haptic feedback, using the shared player.
@@ -61,7 +81,7 @@ public extension HapticFeedback {
     }
     
     /**
-     Trigger the haptic feedback, using the shared player.
+     Trigger the feedback, using the shared player.
      */
     func trigger() {
         Self.player.play(self)