+++ RELEASE_NOTES.md
@@ -42,6 +42,8 @@ This means that you will eventually (not now) have less external dependencies.
 * `KeyboardAction` `.isShift` has been renamed to `isShiftAction`.
 * `KeyboardAction` `.isUppercaseShift` has been renamed to `isUppercasedShiftAction`.
 * `KeyboardImageButton` has been deprecated and will be removed in KeyboardKit 7.
+* `SystemAudio` types have been renamed to `AudioFeedback` to make the concept match haptic feedback types.
+* `SystemAudio` `play` has been renamed to `trigger` to make it match the haptic feedback trigger.